Add files via upload
This is the API for registering a new user. It should update the Users table and create a new Contacts table specific to that user. It connects to the code.js function createContact(). createContact() needs to be updated to also pass in a password value as stated in the API code if not done already.